Refactor CircuitQNN tests (#352)
* refactor tests

* correct docstrings

* code review
